Extract the file
